[aws|iam| make not yet implemented mock errors appear as pending tests

+++ tests/aws/requests/iam/access_key_tests.rb
@@ -1,6 +1,8 @@
 Shindo.tests('AWS::IAM | access key requests', ['aws']) do
 
-  AWS[:iam].create_user('fog_access_key_tests')
+  unless Fog.mocking?
+    AWS[:iam].create_user('fog_access_key_tests')
+  end
 
   tests('success') do
 
@@ -15,6 +17,7 @@ Shindo.tests('AWS::IAM | access key requests', ['aws']) do
     }
 
     tests("#create_access_key('UserName' => 'fog_access_key_tests')").formats(@access_key_format) do
+      pending if Fog.mocking?
       data = AWS[:iam].create_access_key('UserName' => 'fog_access_key_tests').body
       @access_key_id = data['AccessKey']['AccessKeyId']
       data
@@ -30,14 +33,17 @@ Shindo.tests('AWS::IAM | access key requests', ['aws']) do
     }
 
     tests("#list_access_keys('Username' => 'fog_access_key_tests')").formats(@access_keys_format) do
+      pending if Fog.mocking?
       AWS[:iam].list_access_keys('UserName' => 'fog_access_key_tests').body
     end
 
     tests("#update_access_key('#{@access_key_id}', 'Inactive', 'UserName' => 'fog_access_key_tests')").formats(AWS::IAM::Formats::BASIC) do
+      pending if Fog.mocking?
       AWS[:iam].update_access_key(@access_key_id, 'Inactive', 'UserName' => 'fog_access_key_tests').body
     end
 
     tests("#delete_access_key('#{@access_key_id}', 'UserName' => 'fog_access_key_tests)").formats(AWS::IAM::Formats::BASIC) do
+      pending if Fog.mocking?
       AWS[:iam].delete_access_key(@access_key_id, 'UserName' => 'fog_access_key_tests').body
     end
 
@@ -47,6 +53,8 @@ Shindo.tests('AWS::IAM | access key requests', ['aws']) do
     test('failing conditions')
   end
 
-  AWS[:iam].delete_user('fog_access_key_tests')
+  unless Fog.mocking?
+    AWS[:iam].delete_user('fog_access_key_tests')
+  end
 
 end
