+++ tests/aws/requests/iam/user_policy_tests.rb
@@ -1,12 +1,15 @@
 Shindo.tests('AWS::IAM | user policy requests', ['aws']) do
 
-  AWS[:iam].create_user('fog_user_policy_tests')
+  unless Fog.mocking?
+    AWS[:iam].create_user('fog_user_policy_tests')
+  end
 
   tests('success') do
 
     @policy = {"Statement" => [{"Effect" => "Allow", "Action" => "*", "Resource" => "*"}]}
 
     tests("#put_user_policy('fog_user_policy_tests', 'fog_policy', #{@policy.inspect})").formats(AWS::IAM::Formats::BASIC) do
+      pending if Fog.mocking?
       AWS[:iam].put_user_policy('fog_user_policy_tests', 'fog_policy', @policy).body
     end
 
@@ -17,10 +20,12 @@ Shindo.tests('AWS::IAM | user policy requests', ['aws']) do
     }
 
     tests("list_user_policies('fog_user_policy_tests')").formats(@user_policies_format) do
+      pending if Fog.mocking?
       AWS[:iam].list_user_policies('fog_user_policy_tests').body
     end
 
     tests("#delete_user_policy('fog_user_policy_tests', 'fog_policy')").formats(AWS::IAM::Formats::BASIC) do
+      pending if Fog.mocking?
       AWS[:iam].delete_user_policy('fog_user_policy_tests', 'fog_policy').body
     end
 
@@ -30,6 +35,8 @@ Shindo.tests('AWS::IAM | user policy requests', ['aws']) do
     test('failing conditions')
   end
 
-  AWS[:iam].delete_user('fog_user_policy_tests')
+  unless Fog.mocking?
+    AWS[:iam].delete_user('fog_user_policy_tests')
+  end
 
 end
